dev@20316

+++ ultpush-core/src/main/kotlin/com/kairlec/ultpush/core/MessageInfo.kt
@@ -14,6 +14,6 @@ interface MessageInfo {
     /**
      * 消息创建时间
      */
-    val createTime: Int
+    val createTime: Long
 
 }
